Spring cleaning: add simple background
I've added new animations although most of them aren't reachable yet. I've lso tried to make it easier to update the animations, but it seems I've broken the Z sorting (so making this commit before trying to fix that just so that I have something to revert to if it breaks things even more). And I also added a very simple background made of three rectangles, one of which I remove on the results screen because otherwise it would be too bright to see the text in front of it. I also changed the endscreen text to make more sense. I might want to change the Danish version though.